+++ src/vm.c
@@ -44,7 +44,7 @@ stack_init(mrb_state *mrb)
 static void
 stack_extend(mrb_state *mrb, int room, int keep)
 {
-  size_t size, off;
+  int size, off;
 
   if (mrb->stack + room > mrb->stend) {
     size = mrb->stend - mrb->stbase;
@@ -1370,7 +1370,7 @@ mrb_run(mrb_state *mrb, struct RProc *proc, mrb_value self)
       }
       else {
         struct RArray *ary = mrb_ary_ptr(v);
-        size_t len = ary->len;
+        int len = ary->len;
         int i;
 
         if (len > pre + post) {
